+++ BGAnimations/ScreenVOLT26PerformanceSetup background.lua
@@ -1,0 +1,5 @@
+return Def.Quad{
+	InitCommand=function(self)
+		self:FullScreen():diffuse(color("#090909"))
+	end,
+}